refactor DaysCalendar & use DataFetcher
